create packages for activities, fragments & adapters and move all classes to the appropriate package
