+++ app/src/main/java/com/doplgangr/secrecy/activities/FileImportActivity.java
@@ -1,4 +1,4 @@
-package com.doplgangr.secrecy.views;
+package com.doplgangr.secrecy.activities;
 
 import android.app.Fragment;
 import android.app.FragmentManager;
@@ -21,6 +21,8 @@ import com.doplgangr.secrecy.events.AddingFileEvent;
 import com.doplgangr.secrecy.filesystem.encryption.Vault;
 import com.doplgangr.secrecy.filesystem.encryption.VaultHolder;
 import com.doplgangr.secrecy.jobs.AddFileJob;
+import com.doplgangr.secrecy.fragments.FileImportFragment;
+import com.doplgangr.secrecy.adapters.VaultsListFragment;
 
 import java.util.ArrayList;
