+++ app/src/main/java/com/doplgangr/secrecy/fragments/FileViewer.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package com.doplgangr.secrecy.views;
+package com.doplgangr.secrecy.fragments;
 
 import android.app.Fragment;
 import android.content.DialogInterface;
@@ -38,12 +38,14 @@ import com.doplgangr.secrecy.Config;
 import com.doplgangr.secrecy.CustomApp;
 import com.doplgangr.secrecy.R;
 import com.doplgangr.secrecy.Util;
+import com.doplgangr.secrecy.activities.FilesActivity;
 import com.doplgangr.secrecy.filesystem.CryptStateListener;
 import com.doplgangr.secrecy.filesystem.OurFileProvider;
 import com.doplgangr.secrecy.filesystem.Storage;
 import com.doplgangr.secrecy.filesystem.encryption.Vault;
 import com.doplgangr.secrecy.filesystem.files.EncryptedFile;
 import com.doplgangr.secrecy.filesystem.files.SecrecyFile;
+import com.doplgangr.secrecy.fragments.FilesListFragment;
 import com.doplgangr.secrecy.jobs.AddFileJob;
 
 import java.io.File;
