+++ app/src/main/java/com/doplgangr/secrecy/views/dummyviews/LauncherActivity.java
@@ -23,7 +23,7 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.ActionBarActivity;
 
-import com.doplgangr.secrecy.views.MainActivity;
+import com.doplgangr.secrecy.activities.MainActivity;
 
 public class LauncherActivity extends ActionBarActivity {
