+++ app/src/main/res/layout/activity_file_viewer.xml
@@ -25,4 +25,4 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.doplgangr.secrecy.views.FileViewer"/>
+    tools:context="com.doplgangr.secrecy.fragments.FileViewer"/>
