+++ app/src/main/res/menu/file_grid.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.doplgangr.secrecy.views.FilesListFragment">
+    tools:context="com.doplgangr.secrecy.fragments.FilesListFragment">
 
     <item
         android:id="@+id/action_add_file"
