+++ app/src/main/res/menu/main.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.doplgangr.secrecy.views.MainActivity">
+    tools:context="com.doplgangr.secrecy.activities.MainActivity">
 
     <item
         android:id="@+id/action_settings"
